reorganized components and renamed some as well. added type in to foreignData.
