+++ docs/.vuepress/theme/global-components/Documentation.vue
@@ -78,7 +78,7 @@
 
 <script>
 export default {
-  name: 'GenerateTable',
+  name: 'Documentation',
   data() {
     return {
       schema: undefined,
