+++ docs/.vuepress/theme/global-components/DownloadDecks.vue
@@ -40,7 +40,7 @@
 
 <script>
 export default {
-  name: 'GenerateDecksDownloads',
+  name: 'DecksDownloads',
   data() {
     return {
       defaultDecks: null,
