+++ docs/files/all-cards/README.md
@@ -26,4 +26,4 @@ Every card as a unique "atomic" object. May also be outputted as every card in a
 
 ### Data Structure
 
-<GenerateTable/>
+<Documentation/>
