+++ docs/structures/card/README.md
@@ -27,4 +27,4 @@ The card structure describes the properties and values of a single card.
 
 ### Data Structure
 
-<GenerateTable/>
+<Documentation/>
