+++ docs/structures/set/README.md
@@ -26,4 +26,4 @@ The set structure describes the properties and values of a single set.
 
 ### Data Structure
 
-<GenerateTable/>
+<Documentation/>
